+++ CHANGELOG.md
@@ -4,7 +4,18 @@ CloudCompare Version History
 v2.11 (Anoia) - (in development)
 ----------------------
 
+- New features
+  - 3 methods to quickly change the coordinate system of one or several entities
+    - Tools > Registration > Move bounding-box center to origin
+    - Tools > Registration > Move bounding-box min corner to origin
+    - Tools > Registration > Move bounding-box max corner to origin
+
 - Improvements
+  - Clipping box tool:
+    - option to select the extracted contour type (LOWER, UPPER or FULL)
+	  - The 'up' direction is always Z for slices normal to X or Y (the local X, Y or Z directions of the active cross-section)
+	  - The 'up' direction is X for slices normal to Z
+	- all parameters should now be properly remembered from one call to the other (during the same session)
   - Command line tool:
     - The C2M_DIST command (Cloud-to-Mesh distances) can now be called with 2 meshes as input.
         In this case the first mesh vertices are used as compared cloud.
@@ -78,6 +89,7 @@ v2.11 (Anoia) - (in development)
 	- Trace Polyline tool: when changing the OpenGL camera position while tracing a polyline AND using oversampling, strange spikes could appear
 	- the Unroll dialog was not enabling all the apex coordinate fields after switching from Cylinder to Cone mode
 	- the Clipping-box tool 'edit' dialog would sometimes move the box in an unepected way when opening and closing it without making any change
+	- M3C2: the 'subsampling' option was not properly restored when loading the parameters from a file (if 'SubsampleEnabled = false')
 
 v2.10.3 (Zephyrus) - 13/06/2019
 ----------------------